GUARD-910 Since transaction_id (in an Order) can exceed the largest 32-bit int, converted it and other fields that might get large to 64-bit long [v.1.1.0]

+++ src/EtsyAccess/Models/Receipt.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using EtsyAccess.Shared;
 using Newtonsoft.Json;
 
@@ -15,7 +13,7 @@ namespace EtsyAccess.Models
 		/// The receipt's numeric ID
 		/// </summary>
 		[JsonProperty("receipt_id")]
-		public int Id { get; set; }
+		public long Id { get; set; }
 		/// <summary>
 		/// The enum of the order type this receipt is associated with.
 		/// </summary>
@@ -25,7 +23,7 @@ namespace EtsyAccess.Models
 		/// The numeric ID of the order this receipt is associated with.
 		/// </summary>
 		[JsonProperty("order_id")]
-		public int OrderId { get; set; }
+		public long OrderId { get; set; }
 		/// <summary>
 		/// Creation time, in epoch seconds.
 		/// </summary>
